Add WIP AR Notice generator script and resources
Adds dummy data, an R Markdown script, and a Word document formatting template to allow a user to generate an Application Received Public Notice Word document from the extracted .txt data of a completed Groundwater Discharge WMA4 Permit Application fillable pdf. This proof of concept will be tweaked and expanded in the future to allow for generation of multiple permitting documents from a Shiny app webpage.